requirements.txt
absl-py==0.9.0
asgiref==3.2.10      
astroid==2.4.2       
astunparse==1.6.3    
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.1    
certifi==2020.6.20   
chardet==3.0.4
colorama==0.4.3
Django==3.0.8
gast==0.3.3
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
Keras==2.4.3
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
numpy==1.19.0
oauthlib==3.1.0
opt-einsum==3.2.1
Pillow==7.2.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.5.3
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0